# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "fitai.coach Workout API",
    "description": "API for managing workouts, exercises, and training plans.\n\n### Coverage\nThis document is generated from the Zod schemas in `shared/openapi.ts` and\ncurrently describes the workout CRUD and preferences endpoints only. The\nremaining `/api/v1` routes (plans, timeline, analytics, coach, nutrition,\nintegrations, …) are not yet registered here; `server/routes/` is their\nsource of truth.\n\n### Authentication\nThis API is protected using JWT Bearer authentication provided by Clerk.\nTo authenticate your requests:\n1. Obtain a valid JWT session token for your user from Clerk. If you are building a frontend integration, you can use Clerk's `getToken()` method from their SDKs.\n2. Include the token in the `Authorization` header of your HTTP requests as follows:\n   `Authorization: Bearer <your-clerk-jwt-token>`\n\nClick the **Authorize** button below to enter your JWT token for use within this Swagger UI."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Clerk JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "WorkoutId": {
        "type": "string",
        "example": "123e4567-e89b-12d3-a456-426614174000",
        "description": "The UUID of the workout log"
      }
    },
    "parameters": {
      "WorkoutId": {
        "schema": {
          "$ref": "#/components/schemas/WorkoutId"
        },
        "required": true,
        "description": "The UUID of the workout log",
        "name": "id",
        "in": "path"
      }
    }
  },
  "paths": {
    "/api/v1/workouts": {
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Create a new workout log",
        "description": "Creates a new workout log with optional detailed exercise sets (the `exercises` array).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "date": "2026-03-29",
                "focus": "conditioning",
                "mainWorkout": "3 rounds: Sled Push 50m, Burpee Broad Jumps x10, 1km Row",
                "accessory": "Core circuit: 3x20 sit-ups, 3x30s plank",
                "notes": "Hyrox simulation - focus on transitions",
                "duration": 55,
                "rpe": 7,
                "source": "manual",
                "exercises": [
                  {
                    "exerciseName": "Sled Push",
                    "category": "conditioning",
                    "numSets": 3,
                    "sets": [
                      {
                        "setNumber": 1,
                        "distance": 50,
                        "time": 0.7,
                        "notes": "Smooth pace — 42s"
                      },
                      {
                        "setNumber": 2,
                        "distance": 50,
                        "time": 0.75
                      },
                      {
                        "setNumber": 3,
                        "distance": 50,
                        "time": 0.8
                      }
                    ]
                  },
                  {
                    "exerciseName": "Burpee Broad Jumps",
                    "category": "functional",
                    "numSets": 2,
                    "reps": 10,
                    "sets": [
                      {
                        "setNumber": 1,
                        "reps": 10,
                        "notes": "Full extension"
                      },
                      {
                        "setNumber": 2,
                        "reps": 10
                      }
                    ]
                  },
                  {
                    "exerciseName": "Rowing",
                    "category": "conditioning",
                    "numSets": 1,
                    "distance": 1000,
                    "time": 3.5,
                    "sets": [
                      {
                        "setNumber": 1,
                        "distance": 1000,
                        "time": 3.5
                      }
                    ]
                  }
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created workout log with its exercise sets",
            "content": {
              "application/json": {
                "example": {
                  "id": "0b6c5a1e-4f3d-4c2b-9a8e-7d6f5e4c3b2a",
                  "userId": "user_2xY9zQ",
                  "date": "2026-03-29",
                  "focus": "conditioning",
                  "mainWorkout": "3 rounds: Sled Push 50m, Burpee Broad Jumps x10, 1km Row",
                  "accessory": "Core circuit: 3x20 sit-ups, 3x30s plank",
                  "notes": "Hyrox simulation - focus on transitions",
                  "duration": 55,
                  "rpe": 7,
                  "source": "manual",
                  "planDayId": null,
                  "stravaActivityId": null,
                  "calories": null,
                  "distanceMeters": null,
                  "elevationGain": null,
                  "avgHeartrate": null,
                  "maxHeartrate": null,
                  "avgSpeed": null,
                  "maxSpeed": null,
                  "avgCadence": null,
                  "avgWatts": null,
                  "sufferScore": null,
                  "exerciseSets": []
                },
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogWithSets"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get all workout logs",
        "description": "Retrieves a list of all workout logs for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of workout logs",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "0b6c5a1e-4f3d-4c2b-9a8e-7d6f5e4c3b2a",
                    "userId": "user_2xY9zQ",
                    "date": "2026-03-29",
                    "focus": "conditioning",
                    "mainWorkout": "3 rounds: Sled Push 50m, Burpee Broad Jumps x10, 1km Row",
                    "accessory": "Core circuit: 3x20 sit-ups, 3x30s plank",
                    "notes": "Hyrox simulation - focus on transitions",
                    "duration": 55,
                    "rpe": 7,
                    "source": "manual",
                    "planDayId": null,
                    "stravaActivityId": null,
                    "calories": null,
                    "distanceMeters": null,
                    "elevationGain": null,
                    "avgHeartrate": null,
                    "maxHeartrate": null,
                    "avgSpeed": null,
                    "maxSpeed": null,
                    "avgCadence": null,
                    "avgWatts": null,
                    "sufferScore": null,
                    "exerciseSets": []
                  }
                ],
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLog"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/workouts/{id}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get a specific workout log",
        "description": "Retrieves a specific workout log by its ID for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkoutId"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested workout log with its exercise sets, structure blocks and heart-rate RPE suggestion",
            "content": {
              "application/json": {
                "example": {
                  "id": "0b6c5a1e-4f3d-4c2b-9a8e-7d6f5e4c3b2a",
                  "userId": "user_2xY9zQ",
                  "date": "2026-03-29",
                  "focus": "conditioning",
                  "mainWorkout": "3 rounds: Sled Push 50m, Burpee Broad Jumps x10, 1km Row",
                  "accessory": "Core circuit: 3x20 sit-ups, 3x30s plank",
                  "notes": "Hyrox simulation - focus on transitions",
                  "duration": 55,
                  "rpe": 7,
                  "source": "manual",
                  "planDayId": null,
                  "stravaActivityId": null,
                  "calories": null,
                  "distanceMeters": null,
                  "elevationGain": null,
                  "avgHeartrate": null,
                  "maxHeartrate": null,
                  "avgSpeed": null,
                  "maxSpeed": null,
                  "avgCadence": null,
                  "avgWatts": null,
                  "sufferScore": null,
                  "exerciseSets": [],
                  "structureBlocks": [],
                  "suggestedRpe": null
                },
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workout not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Workouts"
        ],
        "summary": "Update a specific workout log",
        "description": "Updates an existing workout log by its ID. You can optionally include an `exercises` array to replace the existing exercise sets.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkoutId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "duration": 52,
                "rpe": 8,
                "notes": "Felt stronger on the sled push rounds today",
                "exercises": [
                  {
                    "exerciseName": "Sled Push",
                    "category": "conditioning",
                    "numSets": 3,
                    "sets": [
                      {
                        "setNumber": 1,
                        "distance": 50,
                        "time": 0.7,
                        "notes": "Smooth pace — 42s"
                      },
                      {
                        "setNumber": 2,
                        "distance": 50,
                        "time": 0.75
                      },
                      {
                        "setNumber": 3,
                        "distance": 50,
                        "time": 0.8
                      }
                    ]
                  },
                  {
                    "exerciseName": "Rowing",
                    "category": "conditioning",
                    "numSets": 1,
                    "distance": 1000,
                    "time": 3.5,
                    "sets": [
                      {
                        "setNumber": 1,
                        "distance": 1000,
                        "time": 3.5
                      }
                    ]
                  }
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated workout log with its exercise sets",
            "content": {
              "application/json": {
                "example": {
                  "id": "0b6c5a1e-4f3d-4c2b-9a8e-7d6f5e4c3b2a",
                  "userId": "user_2xY9zQ",
                  "date": "2026-03-29",
                  "focus": "conditioning",
                  "mainWorkout": "3 rounds: Sled Push 50m, Burpee Broad Jumps x10, 1km Row",
                  "accessory": "Core circuit: 3x20 sit-ups, 3x30s plank",
                  "notes": "Hyrox simulation - focus on transitions",
                  "duration": 55,
                  "rpe": 7,
                  "source": "manual",
                  "planDayId": null,
                  "stravaActivityId": null,
                  "calories": null,
                  "distanceMeters": null,
                  "elevationGain": null,
                  "avgHeartrate": null,
                  "maxHeartrate": null,
                  "avgSpeed": null,
                  "maxSpeed": null,
                  "avgCadence": null,
                  "avgWatts": null,
                  "sufferScore": null,
                  "exerciseSets": []
                },
                "schema": {
                  "$ref": "#/components/schemas/WorkoutLogWithSets"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workout not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Delete a specific workout log",
        "description": "Deletes an existing workout log by its ID for the authenticated user. The log is kept in the recycle bin for 90 days and can be restored.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkoutId"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout log deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "recycleBinItemId": {
                      "type": "string",
                      "description": "Recycle-bin item holding the deleted workout; POST /api/v1/recycle-bin/{id}/restore undoes the delete for 90 days",
                      "example": "8f3c2a1e-5b7d-4c9e-a1f2-3d4e5f6a7b8c"
                    }
                  },
                  "required": [
                    "success",
                    "recycleBinItemId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Workout not found"
          }
        }
      }
    },
...
  }
}