# server/routes/workouts/workoutsCrud.routes.ts
import {
  type AddExerciseSetBody,
  addExerciseSetBodySchema,
  insertWorkoutLogSchema,
  type PatchExerciseSetBody,
  patchExerciseSetBodySchema,
  structureBlockScoreSchema,
} from "@shared/schema";
import { type Request, type Response, Router } from "express";
import { z } from "zod";

import { isAuthenticated } from "../../clerkAuth";
import { DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT } from "../../constants";
import { asyncHandler, parsePagination, rateLimiter, sendNotFound, validateBody, validateQuery } from "../../routeUtils";
import { invalidateAnalyticsCachesForUser } from "../../services/analyticsRouteCache";
import {
  BULK_DELETE_WORKOUTS_NOT_FOUND,
  bulkDeleteWorkouts,
  isBulkDeleteWorkoutsNotFoundError,
} from "../../services/bulkDeleteWorkouts";
import { combineWorkouts } from "../../services/combineWorkouts";
import { loadSuggestedRpe } from "../../services/rpeSuggestion";
import { requestSessionStreamForLog } from "../../services/sessionStreamHooks";
import { deriveMissingWorkoutSetsFromStructure, updateWorkoutStructureBlockScore } from "../../services/workoutService";
import { assignWorkoutPlanDayUseCase, createWorkout, updateWorkoutUseCase } from "../../services/workoutUseCases";
import { storage } from "../../storage";
import { getUserId } from "../../types";
import { createMutateExerciseSetUseCase } from "../../usecases/workouts/mutateExerciseSet.usecase";
import { protectedDelete, protectedPatch, protectedPost } from "../_helpers/protectedRouteBuilder";
import { rejectTextOnlyWriteIfNeeded } from "../structuredWriteGuard";
import { assignWorkoutPlanDaySchema, createCustomExerciseSchema, createWorkoutRouteSchema, updateWorkoutRouteSchema } from "./shared";

const patchExerciseSetSchema = patchExerciseSetBodySchema;
const addExerciseSetSchema = addExerciseSetBodySchema;
const WORKOUT_NOT_FOUND = "Workout not found";
const EXERCISE_SET_NOT_FOUND = "Exercise set not found";

const workoutSetUseCase = createMutateExerciseSetUseCase({
  updateSet: (owner, setId, body, userId) => storage.workouts.mutateExerciseSetUpdate(owner, setId, body, userId),
  addSet: (owner, body, userId) => storage.workouts.mutateExerciseSetAdd(owner, body, userId),
  deleteSet: (owner, setId, userId) => storage.workouts.mutateExerciseSetDelete(owner, setId, userId),
  getUnitPreferences: async (userId) => {
    const user = await storage.users.getUser(userId);
    return { weightUnit: user?.weightUnit, distanceUnit: user?.distanceUnit };
  },
});

// How many past sessions of one exercise to return. The client asks for a
// handful to render "last time"; the ceiling keeps a hand-crafted request
// from pulling an athlete's entire history of a staple lift.
const exerciseHistoryQuerySchema = z.object({
  sessions: z.coerce.number().int().min(1).max(20).optional(),
});

const combineWorkoutsSchema = z.object({
  newWorkout: insertWorkoutLogSchema,
  deleteWorkoutIds: z.array(z.string().min(1)).min(1).max(10),
  skipPlanDayIds: z.array(z.string().min(1)).max(10).optional(),
});

const MAX_BULK_DELETE_TARGETS = 100;
const bulkDeleteWorkoutsSchema = z.object({
  workoutLogIds: z.array(z.string().min(1)).max(MAX_BULK_DELETE_TARGETS).default([]),
  planDayIds: z.array(z.string().min(1)).max(MAX_BULK_DELETE_TARGETS).default([]),
}).superRefine((payload, ctx) => {
  const total = payload.workoutLogIds.length + payload.planDayIds.length;
  if (total === 0) {
    ctx.addIssue({
      code: z.ZodIssueCode.custom,
      message: "Select at least one workout to delete",
    });
  }
  if (total > MAX_BULK_DELETE_TARGETS) {
    ctx.addIssue({
      code: z.ZodIssueCode.custom,
      message: `Select no more than ${MAX_BULK_DELETE_TARGETS} workouts at once`,
    });
  }
});

const updateBlockScoreBodySchema = z.object({
  score: structureBlockScoreSchema.nullable(),
});

export function uniqueIds(ids: readonly string[]): string[] {
  return Array.from(new Set(ids));
}

export function registerWorkoutCrudRoutes(router: Router): void {
  router.get("/api/v1/custom-exercises", isAuthenticated, rateLimiter("customExercise", 60), asyncHandler(async (req: Request, res: Response) => {
    const userId = getUserId(req);
    const exercises = await storage.users.getCustomExercises(userId);
    res.json(exercises);
  }));

  protectedPost(router, "/api/v1/custom-exercises", { limiter: rateLimiter("customExercise", 20), middleware: [validateBody(createCustomExerciseSchema)] }, async (req: Request, res: Response) => {
    const userId = getUserId(req);
    const { name, category } = req.body as { name: string; category?: string };
    const exercise = await storage.users.upsertCustomExercise({ userId, name: name.trim(), category: category || "conditioning" });
    res.json(exercise);
  });

  router.get("/api/v1/workouts", isAuthenticated, rateLimiter("workoutList", 60), asyncHandler(async (req: Request<Record<string, never>, Record<string, never>, Record<string, never>, { limit?: string; offset?: string }>, res: Response) => {
    const userId = getUserId(req);
    const pagination = parsePagination(req.query, res, { defaultLimit: DEFAULT_PAGE_LIMIT, maxLimit: MAX_PAGE_LIMIT });
    if (!pagination) {
      return;
    }
    const logs = await storage.workouts.listWorkoutLogs(userId, pagination.limit, pagination.offset);
    res.json(logs);
  }));

  router.get("/api/v1/workouts/latest", isAuthenticated, rateLimiter("workout", 60), asyncHandler(async (req: Request, res: Response) => {
    const userId = getUserId(req);
    const [latest] = await storage.workouts.listWorkoutLogs(userId, 1);
    if (!latest) {
      return sendNotFound(res, "No workouts found");
    }
    // ⚡ Bolt Performance Optimization:
    // These two reads hit different tables (exercise_sets, workout_structure_blocks)
    // for the same workoutLogId with no ordering dependency between them, so running
    // them sequentially just doubles the round-trip latency for no reason. Promise.all
    // halves the wait on this hot "latest workout" read path.
    const [exerciseSets, structureBlocks] = await Promise.all([
      storage.workouts.getExerciseSetsByWorkoutLog(latest.id),
      storage.workouts.getWorkoutStructureByWorkoutLog(latest.id),
    ]);
    res.json({ ...latest, exerciseSets, structureBlocks });
  }));

  protectedPatch(router, "/api/v1/workouts/:id/sets/:setId", { limiter: rateLimiter("workoutSet", 120), middleware: [validateBody(patchExerciseSetSchema)] }, async (req: Request<{ id: string; setId: string }, Record<string, never>, PatchExerciseSetBody>, res: Response) => {
    const updated = await workoutSetUseCase.updateSet({ kind: "workoutLog", ownerId: req.params.id }, req.params.setId, req.body, getUserId(req));
    if (!updated) {
      return sendNotFound(res, EXERCISE_SET_NOT_FOUND);
    }
    res.json(updated);
  });

  protectedPost(router, "/api/v1/workouts/:id/sets", { limiter: rateLimiter("workoutSet", 60), middleware: [validateBody(addExerciseSetSchema)] }, async (req: Request<{ id: string }, Record<string, never>, AddExerciseSetBody>, res: Response) => {
    const created = await workoutSetUseCase.addSet({ kind: "workoutLog", ownerId: req.params.id }, req.body, getUserId(req));
    if (!created) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    res.status(201).json(created);
  });

  protectedDelete(router, "/api/v1/workouts/:id/sets/:setId", { limiter: rateLimiter("workoutSet", 60) }, async (req: Request<{ id: string; setId: string }>, res: Response) => {
    const deleted = await workoutSetUseCase.deleteSet({ kind: "workoutLog", ownerId: req.params.id }, req.params.setId, getUserId(req));
    if (!deleted) {
      return sendNotFound(res, EXERCISE_SET_NOT_FOUND);
    }
    res.json({ success: true });
  });

  router.get("/api/v1/workouts/:id/history", isAuthenticated, rateLimiter("workoutHistory", 60), asyncHandler(async (req: Request<{ id: string }>, res: Response) => {
    const stats = await storage.workouts.getWorkoutHistoryStats(req.params.id, getUserId(req));
    if (!stats) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    res.json(stats);
  }));

  protectedPost(router, "/api/v1/workouts/:id/seed-from-plan", { limiter: rateLimiter("workoutSet", 20) }, async (req: Request<{ id: string }>, res: Response) => {
    const userId = getUserId(req);
    const seeded = await storage.workouts.seedExerciseSetsFromPlanDay(req.params.id, userId);
    // D10 (CODEBASE_ANALYSIS_2026-10-03): every workout write drops the
    // athlete's cached analytics slices before answering, so the client's
    // refetch reads the write instead of the pre-write cache.
    if (seeded > 0) invalidateAnalyticsCachesForUser(userId);
    res.json({ seededCount: seeded });
  });

  protectedPatch(router, "/api/v1/workouts/:id/structure-blocks/:blockId/score", { limiter: rateLimiter("workout", 40), middleware: [validateBody(updateBlockScoreBodySchema)] }, async (req: Request<{ id: string; blockId: string }, unknown, z.infer<typeof updateBlockScoreBodySchema>>, res: Response) => {
    const structureBlocks = await updateWorkoutStructureBlockScore(req.params.id, req.params.blockId, getUserId(req), req.body.score);
    if (!structureBlocks) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    res.json({ structureBlocks });
  });

  router.get("/api/v1/workouts/:id", isAuthenticated, rateLimiter("workout", 60), asyncHandler(async (req: Request<{ id: string }>, res: Response) => {
    const userId = getUserId(req);
    const log = await storage.workouts.getWorkoutLog(req.params.id, userId);
    if (!log) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    // ⚡ Bolt Performance Optimization: same independent-reads pattern as
    // GET /api/v1/workouts/latest above — parallelize instead of awaiting sequentially.
    const [initialSets, initialBlocks, suggestedRpe] = await Promise.all([
      storage.workouts.getExerciseSetsByWorkoutLog(log.id),
      storage.workouts.getWorkoutStructureByWorkoutLog(log.id),
      // The heart-rate RPE the review sheet offers for the athlete to confirm;
      // never saved on its own (services/rpeSuggestion.ts).
      loadSuggestedRpe(log, userId),
    ]);
    let exerciseSets = initialSets;
    let structureBlocks = initialBlocks;
    if (exerciseSets.length === 0 && structureBlocks.length > 0) {
      await deriveMissingWorkoutSetsFromStructure(log.id, userId);
      [exerciseSets, structureBlocks] = await Promise.all([
        storage.workouts.getExerciseSetsByWorkoutLog(log.id),
        storage.workouts.getWorkoutStructureByWorkoutLog(log.id),
      ]);
    }
    res.json({ ...log, exerciseSets, structureBlocks, suggestedRpe });
  }));

  protectedPost(router, "/api/v1/workouts", { limiter: rateLimiter("workout", 40), middleware: [validateBody(createWorkoutRouteSchema)] }, async (req: Request<Record<string, never>, unknown, z.infer<typeof createWorkoutRouteSchema>>, res: Response) => {
    if (rejectTextOnlyWriteIfNeeded(req, res, "workout_log")) return;
    // Typed by its validator, not cast, so tsc checks the body against createWorkout's payload. CL15 (CODEBASE_ANALYSIS_2026-10-03)
    const result = await createWorkout({ userId: getUserId(req), payload: req.body });
    res.json(result);
  });

  protectedPatch(router, "/api/v1/workouts/:id", { limiter: rateLimiter("workout", 40), middleware: [validateBody(updateWorkoutRouteSchema)] }, async (req: Request<{ id: string }, unknown, z.infer<typeof updateWorkoutRouteSchema>>, res: Response) => {
    if (rejectTextOnlyWriteIfNeeded(req, res, "workout_log")) return;
    // Typed by its validator, not cast, so tsc checks the body against the use case payload. CL15 (CODEBASE_ANALYSIS_2026-10-03)
    const result = await updateWorkoutUseCase({ userId: getUserId(req), workoutId: req.params.id, payload: req.body });
    if (!result) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    res.json(result);
  });

  protectedPatch(router, "/api/v1/workouts/:id/plan-day", { limiter: rateLimiter("workout", 40), middleware: [validateBody(assignWorkoutPlanDaySchema)] }, async (req: Request<{ id: string }>, res: Response) => {
    const { planDayId } = req.body as z.infer<typeof assignWorkoutPlanDaySchema>;
    const userId = getUserId(req);
    const result = await assignWorkoutPlanDayUseCase({ userId, workoutId: req.params.id, planDayId });
    if (!result) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    await requestSessionStreamForLog(storage, userId, result, "assign");
    res.json(result);
  });

  protectedDelete(router, "/api/v1/workouts/:id", { limiter: rateLimiter("workout", 40) }, async (req: Request<{ id: string }>, res: Response) => {
    const userId = getUserId(req);
    const deleted = await storage.workouts.deleteWorkoutLog(req.params.id, userId);
    if (!deleted) {
      return sendNotFound(res, WORKOUT_NOT_FOUND);
    }
    invalidateAnalyticsCachesForUser(userId); // D10
    res.json({ success: true, recycleBinItemId: deleted.recycleBinItemId });
  });

  protectedPost(router, "/api/v1/workouts/bulk-delete", { limiter: rateLimiter("workoutBulkDelete", 20), middleware: [validateBody(bulkDeleteWorkoutsSchema)] }, async (req: Request, res: Response) => {
    const userId = getUserId(req);
    const body = req.body as z.infer<typeof bulkDeleteWorkoutsSchema>;
    const workoutLogIds = uniqueIds(body.workoutLogIds);
    const planDayIds = uniqueIds(body.planDayIds);

    try {
      const result = await bulkDeleteWorkouts({ userId, workoutLogIds, planDayIds });
      invalidateAnalyticsCachesForUser(userId); // D10
      res.json(result);
    } catch (error) {
      if (isBulkDeleteWorkoutsNotFoundError(error)) {
        return sendNotFound(res, BULK_DELETE_WORKOUTS_NOT_FOUND);
      }
      throw error;
    }
  });

  protectedPost(router, "/api/v1/workouts/combine", { limiter: rateLimiter("workout", 10), middleware: [validateBody(combineWorkoutsSchema)] }, async (req: Request, res: Response) => {
    const userId = getUserId(req);
    const { newWorkout, deleteWorkoutIds, skipPlanDayIds } = req.body as z.infer<typeof combineWorkoutsSchema>;
    const result = await combineWorkouts({ userId, newWorkout, deleteWorkoutIds, skipPlanDayIds });
    invalidateAnalyticsCachesForUser(userId); // D10
    res.status(201).json(result);
  });

  // Its own rate-limit bucket, not the one shared with /workouts/:id/history:
  // opening a session fires one of these per distinct exercise, so a 10-lift
  // day would eat a sixth of a 60/min budget in a single screen.
  router.get("/api/v1/exercises/:exerciseName/history", isAuthenticated, rateLimiter("exerciseHistory", 120), validateQuery(exerciseHistoryQuerySchema), asyncHandler(async (req: Request<{ exerciseName: string }>, res: Response) => {
    const { sessions } = req.query as z.infer<typeof exerciseHistoryQuerySchema>;
    const history = await storage.workouts.getExerciseHistory(getUserId(req), req.params.exerciseName, { sessionLimit: sessions });
    res.json(history);
  }));
}